fuck stuff up
had this idea that it would be better if we only did platform
rendering as fast as we can during vblank then did other things
after vblank, but that extra time really does help get those
drawn, so really this just amounted to a lot of trashing
about on how that was handled.
I'm not sure what went wrong but the gap filling had to
get a few more jmps to deal with the larger in-lined vblank handling
even though this was inlined before and I don't think I really
grew it other than fixing a bug where we needed one more vblank
line.